handle update with auth

+++ Backend-Express/controllers/FotoController.js
@@ -3,6 +3,7 @@ const prisma = new PrismaClient();
 const NotFound = require("../exceptions/NotFound");
 const ValidationError = require("../exceptions/ValidationError");
 const { validationResult } = require("express-validator");
+const AuthError = require("../exceptions/AuthError");
 
 
 async function index(req, res, next) {
@@ -124,9 +125,57 @@ async function store(req, res, next) {
 
 
 async function update(req, res, next) {
+    try {
+        const { id } = req.params;
+        const data = await prisma.foto.findUnique({
+            where: {
+                id: parseInt(id),
+            },
+        });
 
+        if (!data) {
+            const notFound = new NotFound("La foto indicata non è stata trovata.");
+            res.status(notFound.status).json(notFound.resp());
+        }
+
+        if (data.userId !== req.user.id) {
+            const authError = new AuthError("Accesso non autorizzato", 403);
+            res.status(authError.status).json(authError.resp());
+        }
+
+        const file = req.file;
+        const datiInIngresso = req.validatedData;
+        let imageToUpdate;
+
+        if (file) {
+            imageToUpdate = file.filename;
+        } else {
+            // Se non è stato caricato un nuovo file, mantieni l'immagine esistente
+            imageToUpdate = datiInIngresso.image || 'placeholder';
+        }
+
+        const fotoUpdate = await prisma.foto.update({
+            where: {
+                id: parseInt(req.params.id),
+            },
+            data: {
+                name: datiInIngresso.name,
+                description: datiInIngresso.description,
+                visible: datiInIngresso.visible,
+                image: imageToUpdate,
+                categories: {
+                    connect: datiInIngresso.categories?.map((idCategory) => ({
+                        id: +idCategory
+                    }))
+                }
+            }
+        });
+        return res.json({ message: `Aggiornamento di ${data.name} completato con successo` });
+
+    } catch (error) {
+        next("Qualcosa è andato storto durante l'aggiornamento");
+    }
 }
-
 
 async function destroy(req, res, next) {
     try {
@@ -143,13 +192,19 @@ async function destroy(req, res, next) {
 
         }
 
-        await prisma.foto.delete({
-            where: {
-                id: parseInt(id),
-            },
-        });
+        if (data.userId === req.user.id) {
+            await prisma.foto.delete({
+                where: {
+                    id: parseInt(id),
+                },
+            });
 
-        return res.json({ message: `${data.name} è stato eliminato con successo` });
+            return res.json({ message: `${data.name} è stato eliminato con successo` });
+        } else {
+            const authError = new AuthError("Accesso non autorizzato", 403);
+            res.status(authError.status).json(authError.resp());
+        }
+
     } catch (error) {
         next(error);
     }
